Rename `preview2::preview2` to `preview2::host`
