+++ crates/wasi/src/preview2/host/exit.rs
